// src/components/Enums.ts
export enum Colors {
    bgRed = 'bg-red-50 hover:bg-red-200 text-black',
    bgBlue = 'bg-blue-50 hover:bg-blue-200 text-black',
    bgGreen = 'bg-green-50 hover:bg-green-200 text-black',
    bgGray = 'bg-gray-50 hover:bg-gray-200 text-black',
    bgCalc = 'bg-gray-100 px-4 py-4 shadow-[3px_3px_4px_rgba(0,0,0,0.1)]',
    border = 'border rounded-xl border border-gray-300',
    grid = 'grid grid-rows-5 grid-flow-col gap-4',
    shadow = 'shadow-[2px_2px_4px_rgba(0,0,0,0.1)]',
}